requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
asyncio
aiofiles>=23.0.0